Make the message catalog UTF-8 encoding detection slightly more broad
Message catalogs are supposed to be UTF-16BE, but not everyone follows
that. We already had a workaround that assumed message catalogs would
start with, well, a message. Iberia breaks that assumption by adding
what looks like a comment first (which I'm not sure the catalog file
format even allows for...).